sync

+++ torrent.go
@@ -131,7 +131,7 @@ func (c *Client) handleBitfield(peer *Peer, torrent *Torrent, payload []byte) {
 func (c *Client) handleRequest(peer *Peer, torrent *Torrent, payload []byte) {
 	// fmt.Println("===== HANDLE  REQUEST   =======")
 	indx := binary.BigEndian.Uint32(payload[:4])
-	begin := binary.BigEndian.Uint32(payload[4:8])
+	begin := int64(binary.BigEndian.Uint32(payload[4:8]))
 	length := int64(binary.BigEndian.Uint32(payload[8:12]))
 	fileMap := torrent.PieceMap[indx].FileMap
